Revert "refactor: change name of main.dart to Home.dart"
This reverts commit ad24c771
